CTB-97 Get StockChart data from backend FEATURE
